Partially fixing rate limit issue
addressing #18 trying to limit API calls
Also at first this Issue wasn't clear since "Clients are also limited to one gateway connection per 5 seconds. If you hit this limit, the Gateway will respond with an Opcode 9 Invalid Session." the opcode 9 isnt caught by the lib and technically it is not the API Rate limit (but a "gateway rate limit" which was probably hit) 
Now we should use aloooot less calls as we only build the needed arrays every now and then as newly joined users are unlikely to be protected.

We cannot influence how often we get the "Guildmemberchange" event in which we catch renames however. (Generally each event is a call in the backend) so if it gets too active this will disconnect the bot from the API from discord's side

Needs improovements f.E in line #170 we could query our cache first to prevent the "getmember" api call if we already have the messaging user.