Push Current Running Version
This commit establishes a mirror of the running code in the wild (minus some configuration params, they may differ) 
This is so anyone can audit the Bot thats running. 
Please note that the master branch is the dev branch, and is only Stable on given Tags! (There are many changes on master that may not make it into a Release)